Disallow use of undefined variables in RedirectRule URLs
